Die letzten groeberen Bugs in ccm-sci-publications sollten jetzt behoben sein.
git-svn-id: https://svn.libreccm.org/ccm/trunk@521 8810af33-2d31-482b-a856-94f89814c4df

+++ ccm-sci-publications/src/com/arsdigita/cms/contenttypes/Series.java
@@ -72,7 +72,7 @@ public class Series extends ContentPage {
         DataObject link = add(EDITORS, editor);
         link.set(EDITOR_FROM, from);
         link.set(EDITOR_TO, to);
-        link.set(EDITOR_ORDER, BigDecimal.valueOf(getEditors().size()));
+        link.set(EDITOR_ORDER, Integer.valueOf((int)getEditors().size()));
     }
 
     public void removeEditor(GenericPerson editor) {
